add  login jwt token

+++ server/src/main/java/etu/spb/etu/Internet_news_newspaper/like/service/LikeServiceImpl.java
@@ -2,6 +2,7 @@ package etu.spb.etu.Internet_news_newspaper.like.service;
 
 import etu.spb.etu.Internet_news_newspaper.exception.AlreadyExistException;
 import etu.spb.etu.Internet_news_newspaper.exception.IdNotFoundException;
+import etu.spb.etu.Internet_news_newspaper.exception.NotOwnerException;
 import etu.spb.etu.Internet_news_newspaper.like.Like;
 import etu.spb.etu.Internet_news_newspaper.like.LikeRepository;
 import etu.spb.etu.Internet_news_newspaper.post.model.Post;
@@ -10,6 +11,8 @@ import etu.spb.etu.Internet_news_newspaper.user.UserRepository;
 import etu.spb.etu.Internet_news_newspaper.user.model.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -35,9 +38,15 @@ public class LikeServiceImpl implements LikeService {
     }
 
     @Override
-    public void deleteLike(Long likeId) {
+    public void deleteLike(Long likeId, Long userId) {
         Like like = likeRepository.findById(likeId)
                 .orElseThrow(() -> new IdNotFoundException("Лайк c id = " + likeId + " не найден"));
+
+        if (!Objects.equals(like.getUserId(), userId)) {
+            throw new NotOwnerException("Лайк может удалить только пользователь, поставивший его," +
+                    " пользователь с id = " + userId + " не является пользователем лайка");
+        }
+
         likeRepository.deleteById(likeId);
     }
